+++ src/main/java/ipp/aci/boleia/dominio/parametros/LogicaParametroPrecoMaximo.java
@@ -37,13 +37,13 @@ public class LogicaParametroPrecoMaximo implements ILogicaParametroSistema<Autor
         if(veiculo.isProprio()) {
             Map<Long, BigDecimal> precoMaximoProdutos = obterMapaProdutos(frotaParam.getPrecoMaximoProdutos());
             Map<Long, BigDecimal> precoMaximoCombustiveis = obterMapaCombustiveis(frotaParam.getPrecoMaximoAbastecimentos());
-            Map<Long, Long> quantidadeMaximaProdutosPorUnidade= obterMapaQuantidadePorProduto(frotaParam.getQuantidadeMaximaPorProduto());
+            Map<Long, BigDecimal> quantidadeMaximaProdutosPorUnidade= obterMapaQuantidadePorProduto(frotaParam.getQuantidadeMaximaPorProduto());
             Map<Long, BigDecimal> quantidadeMaximaLitrosCombustivel = obterMapaQuantidadePorAbastecimento(frotaParam.getQuantidadeMaximaPorCombustivel());
             StringBuilder produtosViolados = new StringBuilder();
             for (ItemAutorizacaoPagamento item : autorizacao.getItems()) {
                 BigDecimal precoMaximoPermitido;
                 BigDecimal quantidadeMaximaPermitidaAbastecimento = null;
-                Long quantidadeMaximaPermitidaProduto = null;
+                BigDecimal quantidadeMaximaPermitidaProduto = null;
 
                 if (item.isAbastecimento()) {
                     precoMaximoPermitido = precoMaximoCombustiveis.get(item.getCombustivel().getId());
@@ -52,7 +52,7 @@ public class LogicaParametroPrecoMaximo implements ILogicaParametroSistema<Autor
                     precoMaximoPermitido = precoMaximoProdutos.get(item.getProduto().getId());
                     quantidadeMaximaPermitidaProduto = quantidadeMaximaProdutosPorUnidade.get(item.getProduto().getId());
                 }
-                if ((quantidadeMaximaPermitidaProduto != null && item.getQuantidade().compareTo((BigDecimal.valueOf(quantidadeMaximaPermitidaProduto))) > 0 || quantidadeMaximaPermitidaAbastecimento != null && item.getQuantidade().compareTo((quantidadeMaximaPermitidaAbastecimento)) > 0) ||(precoMaximoPermitido != null && item.getValorUnitario().compareTo(precoMaximoPermitido) > 0)) {
+                if ((quantidadeMaximaPermitidaProduto != null && item.getQuantidade().compareTo(quantidadeMaximaPermitidaProduto) > 0 || quantidadeMaximaPermitidaAbastecimento != null && item.getQuantidade().compareTo((quantidadeMaximaPermitidaAbastecimento)) > 0) ||(precoMaximoPermitido != null && item.getValorUnitario().compareTo(precoMaximoPermitido) > 0)) {
                     resultado.setStatusResultado(StatusExecucaoParametroSistema.ERRO);
                     produtosViolados.append(mensagens.obterMensagem("parametro.sistema.erro.abastecimento.preco.maximo.produto", item.getNome(), UtilitarioFormatacao.formatarDecimal(precoMaximoPermitido), UtilitarioFormatacao.formatarDecimal(item.getValorUnitario())));
                 }
@@ -69,8 +69,8 @@ public class LogicaParametroPrecoMaximo implements ILogicaParametroSistema<Autor
      * @param quantidadePermitidaProdutos
      * @return retorna um mapa de quantidade de produtos que foram obtidos
      */
-    private Map<Long,Long> obterMapaQuantidadePorProduto(List<FrotaParametroSistemaPrecoMaximoProduto> quantidadePermitidaProdutos) {
-        Map<Long,Long> resultado = new HashMap<>();
+    private Map<Long,BigDecimal> obterMapaQuantidadePorProduto(List<FrotaParametroSistemaPrecoMaximoProduto> quantidadePermitidaProdutos) {
+        Map<Long,BigDecimal> resultado = new HashMap<>();
         quantidadePermitidaProdutos.forEach(quantidadeMaxima ->
                 resultado.put(quantidadeMaxima.getProduto().getId(), quantidadeMaxima.getQuantidadePermitida())
         );
